chore(adapters): move Markdown adapter implementation into markdown subdirectory
Move markdown.ts, extract.ts, apply.ts, mdx-rules.ts, mdx-placeholders.ts, mdx-jsx-attributes.ts, mdx-static-data.ts, parser.ts, and traverse.ts from src/ to src/adapters/markdown/. Update adapter.ts to make FileAdapter generic over extract/apply options types. Update index.ts barrel exports to reference new paths. Update README.md file locations to reflect new structure. Add MarkdownExtractOptions and MarkdownApplyOptions interfaces extending base adapter options with
