+++ .changeset/tidy-adapter-layout.md
@@ -1,0 +1,9 @@
+---
+"@cloudflare/polystella": minor
+"@cloudflare/polystella-astro": minor
+"@cloudflare/polystella-core": minor
+"@cloudflare/polystella-adapters": minor
+"@cloudflare/polystella-providers": minor
+---
+
+Colocate the Markdown and MDX adapter implementation and move MDX-specific options out of the generic adapter contracts.
