+++ packages/adapters/src/adapters/markdown/apply.ts
