+++ packages/adapters/src/adapters/markdown/extract.ts
