+++ packages/adapters/src/adapters/markdown/mdx-jsx-attributes.ts
