+++ packages/adapters/src/adapters/markdown/mdx-placeholders.ts
