+++ packages/adapters/src/adapters/markdown/mdx-rules.ts
