+++ packages/adapters/src/adapters/markdown/mdx-static-data.ts
