+++ packages/adapters/src/adapters/markdown/parser.ts
