+++ packages/adapters/src/adapters/markdown/traverse.ts
