+++ packages/adapters/src/index.ts
@@ -1,11 +1,11 @@
 export * from "./adapter.js";
 export * from "./adapters/json.js";
-export * from "./adapters/markdown.js";
+export * from "./adapters/markdown/adapter.js";
+export * from "./adapters/markdown/apply.js";
+export * from "./adapters/markdown/extract.js";
+export * from "./adapters/markdown/mdx-placeholders.js";
+export * from "./adapters/markdown/mdx-rules.js";
+export * from "./adapters/markdown/parser.js";
 export * from "./adapters/toml.js";
 export * from "./adapters/yaml.js";
-export * from "./apply.js";
-export * from "./extract.js";
 export * from "./key-paths.js";
-export * from "./mdx-placeholders.js";
-export * from "./mdx-rules.js";
-export * from "./parser.js";
